Reorg pds record plugins into record service
