+++ packages/pds/src/services/record/index.ts
@@ -2,18 +2,18 @@ import { CID } from 'multiformats/cid'
 import { ValidationError } from '@atproto/lexicon'
 import { AtUri } from '@atproto/uri'
 import * as common from '@atproto/common'
-import { dbLogger as log } from '../logger'
-import Database from '../db'
-import * as Declaration from '../db/records/declaration'
-import * as Post from '../db/records/post'
-import * as Vote from '../db/records/vote'
-import * as Repost from '../db/records/repost'
-import * as Trend from '../db/records/trend'
-import * as Follow from '../db/records/follow'
-import * as Assertion from '../db/records/assertion'
-import * as Confirmation from '../db/records/confirmation'
-import * as Profile from '../db/records/profile'
-import { MessageQueue } from '../stream/types'
+import { dbLogger as log } from '../../logger'
+import Database from '../../db'
+import * as Declaration from './plugins/declaration'
+import * as Post from './plugins/post'
+import * as Vote from './plugins/vote'
+import * as Repost from './plugins/repost'
+import * as Trend from './plugins/trend'
+import * as Follow from './plugins/follow'
+import * as Assertion from './plugins/assertion'
+import * as Confirmation from './plugins/confirmation'
+import * as Profile from './plugins/profile'
+import { MessageQueue } from '../../stream/types'
 
 export class RecordService {
   records: {
