+++ packages/pds/src/services/record/plugins/profile.ts
@@ -1,14 +1,14 @@
 import { Kysely } from 'kysely'
 import { AtUri } from '@atproto/uri'
 import { CID } from 'multiformats/cid'
-import * as Profile from '../../lexicon/types/app/bsky/actor/profile'
-import { Profile as IndexedProfile } from '../tables/profile'
-import * as lex from '../../lexicon/lexicons'
-import { Message } from '../../stream/messages'
-import DatabaseSchema from '../database-schema'
-import RecordProcessor from '../record-processor'
+import * as Profile from '../../../lexicon/types/app/bsky/actor/profile'
+import * as lex from '../../../lexicon/lexicons'
+import { Message } from '../../../stream/messages'
+import DatabaseSchema from '../../../db/database-schema'
+import RecordProcessor from '../processor'
 
 const lexId = lex.ids.AppBskyActorProfile
+type IndexedProfile = DatabaseSchema['profile']
 
 const insertFn = async (
   db: Kysely<DatabaseSchema>,
